==> Requirements-Tracing-main/2.1-preprocessForTfidf.ipynb <==
# Preprocessing 

All of these steps are optional. Finding the right way to preprocess your text data will be accomplished through trial and error. Additional, more targeted preprocessing may be necessary for each subset if you notice abnormalities that might affect your clustering (like ID numbers, specific formatting, etc.) My advice would be to use the Description and any other text data available to you if and when it makes sense to do so. The more text, the better, but only if it provides new and relevant information. In other words, don’t inject noise for the sake of adding more text data. 

Removing stop words (e.g., “the”, “and”, etc).  

You can also create your own list of common words to remove if you find that your clusters are being impacted by words that aren’t really relevant (this will be revealed when you do key word extraction on your clusters). For example, you might find that you get clusters of tickets that all have vague words like “requirements” or “user.” While the language in those clusters is similar, it’s not really accomplishing the task we’re setting out to do, so you may want to consider removing these types of words from the text. You may also consider removing all Components and system names from the text. 

- Removing punctuation and/or digits 

- Lower casing 

- Lemmatizing  

This will convert all words to their base form. For example, it would change “running” to “run” and “better” to “good.” This is more robust than stemming. 

If you’re using Doc2Vec or another algorithm that takes context into consideration, you may want to skip this step. Trial and error will help you make that decision.

# import packages
import pandas as pd
import utils
import pickle

# Import Processed Requirements Data from saved CSV

df = pd.read_csv('dummy_requirements.csv')

# Preprocess Text for NLP evaluation

df = df.astype(str)

# ADD STOP WORDS HERE
my_stop_words = ['the', 'system', 'shall', 'application', 'an', 'a', 'using', 
                 'standard', 'that', 'this']

df['Requirement'] = df.apply(lambda x: utils.preprocess_text(x['Description'], my_stop_words, 
                                                             keep_original=False, display=False, 
                                                             pos_tags=True), 
                                                             axis=1)

df['POS'] = df.apply(lambda x: x['Requirement'][1], axis=1)
df['Requirement'] = df.apply(lambda x: x['Requirement'][0],axis=1)
df

with open("cleaned_dummy_tfidf.pickle", "wb") as pickle_file:
    pickle.dump(df, pickle_file, protocol=pickle.HIGHEST_PROTOCOL)

# View list of Components

# Number of Requirements by Component/Program of Record
components = df.groupby(by='Component/s')
components.size()

# Subgroup by component

groups = components.groups.keys()
grouped_data = {}
for g in groups:
    print(g)
    sub_df = components.get_group(g)
    sub_df = sub_df.astype(str)
    X = sub_df
    grouped_data[g] = X
    print(X)

# Show output
grouped_data

with open("DummyPreproccessedForTfidf.pickle", "wb") as pickle_file:
    pickle.dump(grouped_data, pickle_file, protocol=pickle.HIGHEST_PROTOCOL)

## Legend for Parts of Speech (POS): 

- CC coordinating conjunction
- CD cardinal digit
- DT determiner
- EX existential there (like: “there is” … think of it like “there exists”)
- FW foreign word
- IN preposition/subordinating conjunction
- JJ adjective ‘big’
- JJR adjective, comparative ‘bigger’
- JJS adjective, superlative ‘biggest’
- LS list marker 1)
- MD modal could, will
- NN noun, singular ‘desk’
- NNS noun plural ‘desks’
- NNP proper noun, singular ‘Harrison’
- NNPS proper noun, plural ‘Americans’
- PDT predeterminer ‘all the kids’
- POS possessive ending parent‘s
- PRP personal pronoun I, he, she
- PRPS possessive pronoun my, his, hers
- RB adverb very, silently,
- RBR adverb, comparative better
- RBS adverb, superlative best
- RP particle give up
- TO to go ‘to‘ the store.
- UH interjection errrrrrrrm
- VB verb, base form take
- VBD verb, past tense took
- VBG verb, gerund/present participle taking
- VBN verb, past participle taken
- VBP verb, sing. present, non-3d take
- VBZ verb, 3rd person sing. present takes
- WDT wh-determiner which
- WP wh-pronoun who, what
- WPS possessive wh-pronoun whose
- WRB wh-abverb where, when
