==> Requirements-Tracing-main/4-DetermineOptimalKClusters.ipynb <==
# import packages
import numpy as np
import pandas as pd
import pickle

# Load grouped data by component from saved pickle

# Load grouped data by component from saved pickle

with open("DummyPreproccessedForTfidf.pickle", "rb") as pickle_file:
    grouped_data = pickle.load(pickle_file)
    
    
with open("VectorizationTfidf.pickle", "rb") as pickle_file:
    group_X = pickle.load(pickle_file)

### Load Doc2Vec Pickles ###

# with open("DummyPreproccessedForDoc2Vec.pickle", "rb") as pickle_file:
#     grouped_data = pickle.load(pickle_file)
    
    
# with open("VectorizationDoc2Vec.pickle", "rb") as pickle_file:
#     group_X = pickle.load(pickle_file)

g = 'ALPHA'

# Determine optimal K clusters

### Determining optimal K clusters ###
# Uses elbow method and silhouette coefficient
# This takes a while...
from sklearn.metrics import silhouette_score
from sklearn.cluster import KMeans

silhouette_coefficients = []
Sum_of_squared_distances = []
K = range(2, 100)
kmeans_kwargs = {
         "init": "random",
         "n_init": 10,
         "max_iter": 300,
         "random_state": 42,
     }
for k in K: 
    km = KMeans(n_clusters=k, **kmeans_kwargs)
    km = km.fit(group_X[g])
    Sum_of_squared_distances.append(km.inertia_)
#     print(km.labels_)
    score = silhouette_score(group_X[g], km.labels_)
    silhouette_coefficients.append(score)
    
    
### One method to find the elbow is to ###
### find the point where the slope changes the most (optimize 2nd derivative). ###
dx_sum_of_squared_distances = np.diff(np.asarray(Sum_of_squared_distances))
dx2_sum_of_squared_distances = np.diff(dx_sum_of_squared_distances)
elbow_max = max(dx2_sum_of_squared_distances)
elbow_index = list(dx2_sum_of_squared_distances).index(elbow_max) + 1 + min(K)
print(f'# of Clusters using Elbow Method = {elbow_index}.')
    
import matplotlib.pyplot as plt

plt.plot(K, Sum_of_squared_distances, 'bx-')
plt.xlabel('k')
plt.ylabel('Sum_of_squared_distances')
plt.title('Elbow Method For Optimal k')
plt.show()

plt.style.use("fivethirtyeight")
plt.plot(K, silhouette_coefficients)
plt.xticks(K)
plt.xlabel("Number of Clusters")
plt.ylabel("Silhouette Coefficient")
plt.show()

silhouette_coefficients

max_value = max(silhouette_coefficients)
max_index = silhouette_coefficients.index(max_value)
print(f'# of Clusters using Silhouette Coefficient = {max_index}.')

## Plot Again

import matplotlib.pyplot as plt

plt.plot(K,Sum_of_squared_distances, 'bx-')
plt.axvline(x=elbow_index)
plt.axvline(x=max_index)
plt.xlabel('k')
plt.ylabel('Sum_of_squared_distances')
plt.title('Elbow Method For Optimal k')
plt.show()

plt.style.use("fivethirtyeight")
plt.plot(K, silhouette_coefficients)
plt.axvline(x=elbow_index)
plt.axvline(x=max_index)
plt.xticks(K)
plt.xlabel("Number of Clusters")
plt.ylabel("Silhouette Coefficient")
plt.show()
